ajustes de código
