+++ src/main/java/com/sgb/biblioteca/controller/normal/LivroController.java
@@ -1,6 +1,8 @@
 package com.sgb.biblioteca.controller.normal;
 
 import com.sgb.biblioteca.model.Livro;
+import com.sgb.biblioteca.service.AutorService;
+import com.sgb.biblioteca.service.EditoraService;
 import com.sgb.biblioteca.service.GeneroService;
 import com.sgb.biblioteca.service.LivroService;
 
@@ -13,6 +15,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.ModelAndView;
 
 
 @Controller 
@@ -21,28 +24,29 @@ import org.springframework.web.bind.annotation.RequestMapping;
 public class LivroController {
 
     private LivroService livroService;
+
     private GeneroService generoService;
 
     @GetMapping("/{id}")
-    public String get(@PathVariable Long id, Model model){
+    public ModelAndView get(@PathVariable Long id){
         val livro = livroService.findLivroById(id);
-        model.addAttribute("livro", livro);
 
-        return "livro/get";
+        return new ModelAndView("livro/get")
+            .addObject("livro", livro);
     }
 
     @GetMapping("/new")
-    public String edit(Model model){
+    public ModelAndView edit(){
         val generos = generoService.findAllGeneros();
 
-        model.addAttribute("generos", generos);
-        
-        return "livro/edit";
+        return new ModelAndView("livro/edit")
+            .addObject("generos", generos);
     }
 
     @PostMapping("/new")
     public String post(Livro livro){
         livroService.save(livro);
+        
         return "livro/edit";
     }
 }
