+++ src/main/resources/templates/layout/navbar.html
@@ -21,7 +21,7 @@
                         <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#" role="button" aria-haspopup="true" aria-expanded="false">Livros</a>
                         <div class="dropdown-menu">
                             <a class="dropdown-item" href="#">Listar livros</a>
-                            <a class="dropdown-item" href="#">Cadastrar</a>
+                            <a class="dropdown-item" href="../livro/edit.html" th:href="@{new}">Cadastrar</a>
                         </div>
                     </li>
                     <li class="nav-item dropdown">
@@ -50,7 +50,6 @@
     </nav>
 
 <script src="../static/js/3rd/popper.min.js" th:src="@{/js/3rd/popper.min.js}"></script>
-<script src="../static/js/3rd/jquery.js" th:src="@{/js/3rd/jquery.js}"></script>
 <script src="../static/js/3rd/bootstrap.min.js" th:src="@{/js/3rd/bootstrap.min.js}"></script>
 
 </body>
