--- src/main/resources/templates/teste.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org"
-      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout">
-<head>
-    <meta charset="UTF-8">
-    <title>Cadastro de Livro</title>
-    <link href="../static/css/3rd/bootstrap.min.css"            th:href="@{/css/3rd/bootstrap.min.css}"        rel="stylesheet" media="screen" />
-
-</head>
-<body>
-    <div th:replace="~{ layout/navbar :: navbar}"></div>
-    <h1>Cadastro de Livro</h1>
-    <form method="post" th:action="@{/biblioteca}">
-        <div>
-            <label for="titulo">Título:</label>
-            <input type="text" id="titulo" name="titulo" required>
-        </div>
-        <div>
-            <label for="quantidade">Quantidade:</label>
-            <input type="number" id="quantidade" name="quantidade" required>
-        </div>
-        <div>
-            <input type="submit" class="btn btn-primary" value="Salvar">
-        </div>
-    </form>
-    <script src="../static/js/3rd/bootstrap.min.js" th:src="@{/js/3rd/bootstrap.min.js}"></script>
-</body>
-</html>
